+++ .github/CHANGELOG.md
@@ -107,6 +107,7 @@ Always start an item with "1."; GitHub will number the rendered list automatical
 1. [A380X/RMP] Fix RMP brightness initialisation - @CameronCarlyon (YouBaconMeCrazy)
 1. [A380X] Make flightdeck lighting panel-state and time-of-day aware. Adjust existing values. @CameronCarlyon (YouBaconMeCrazy)
 1. [A32NX/A380X] Corrected off-by-.01 conversion value for kohlsman event - @rthom91 (Randy)
+1. [A32NX/FMS] Limit step descent guidance to -1000fpm - @Jonny23787 (Jonathan), @BlueberryKing (BlueberryKing)
 
 ## 2024.1.0